WIP clang: __ATTR_PROGMEM__ (a bit broken?)
Clang builds with this flag are completely garbled on screen. There
seems to be a distinct garble for the splash and the gameplay, so it
seems the program is running to some extent.

avr-gcc builds with this flag (i.e. this commit) seem to garble the
splash screen (100% white?) but the game itself runs fine.